[ads][CodeHealth] Decouple TestApplicationStateObserver into test/
Move the inline test observer out of the anonymous namespace in
application_state_monitor_unittest.cc into a dedicated
test/test_application_state_observer.h header and add a test_support
source_set in application_state/BUILD.gn.

No behavioral changes.
